Proyecto Final/requirement.txt
mysql-connector==8.0.40